Move InstanceStateRefreshFunc to right file
